Proyectos/tienda-django/requirements/local.txt
asgiref==3.2.7
Django==3.0.5
django-model-utils==4.0.0
djangorestframework==3.11.0
Pillow==9.1.0
psycopg2==2.8.6
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1